Replaced old manager metadata manager
